+++ CHANGELOG.md
@@ -12,7 +12,7 @@
 - [418](https://github.com/vegaprotocol/protos/issues/418) - Add support for cursor based pagination in `Data Node API V2` Trades, Markets and Parties
 - [427](https://github.com/vegaprotocol/protos/issues/418) - Add support for cursor based pagination in `Data Node API V2` Orders
 - [5358](https://github.com/vegaprotocol/vega/issues/5358) - Add previous MVP field to equity shares snapshot.
-- [](https://github.com/vegaprotocol/protos/pull/) -
+- [437](https://github.com/vegaprotocol/protos/issues/437) - Add `UpdateAsset` proto definition.
 
 ### 🐛 Fixes
 - [433](https://github.com/vegaprotocol/protos/pull/433) - remove unnecessary semicolon from vega proto